feat: 음식점 선택지 확인 페이지 구현 (#10)
