+++ src/components/Navbar.tsx
@@ -12,9 +12,9 @@ const HeaderContainer = styled.header`
     display: flex;
     justify-content: space-between;
     align-items: center;
-    padding: 15px 25px;
+    padding: 0px 25px;
     background-color: #ffffff;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    height: 60px;
 `;
 
 const Logo = styled.h1`
